+++ src/api/player/THEOplayer.ts
@@ -284,7 +284,14 @@ export interface THEOplayer extends EventDispatcher<PlayerEventMap> {
   readonly videoHeight: number | undefined;
 
   /**
-   * The API for theolive.
+   * The API for THEOlive.
+   */
+  readonly theoLive: TheoLiveAPI;
+
+  /**
+   * The API for THEOlive.
+   *
+   * @deprecated use {@link THEOplayer.theoLive} instead.
    */
   readonly theolive: TheoLiveAPI;
 }